+++ demo/gene_regulation.h
@@ -26,29 +26,29 @@ inline int Simulate(int argc, const char** argv) {
   // 3. Initialize RegulateGenes module.
   // To add functions to the module use RegulateGenes::AddGene() function.
   // You should pass to the function two variables.
-  // The first is of type  std::function<double(double, double)>.
+  // The first is of type  std::function<float(float, float)>.
   // This is the function by which concentration of the protein will be
   // calculated.
-  // The second is double. This is the initial value for the protein.
+  // The second is float. This is the initial value for the protein.
   RegulateGenes regulate_example;
   regulate_example.AddGene(
-      [](double curr_time, double last_concentration) {
+      [](float curr_time, float last_concentration) {
         return curr_time * last_concentration + 0.2f;
       },
       1);
   regulate_example.AddGene(
-      [](double curr_time, double last_concentration) {
+      [](float curr_time, float last_concentration) {
         return last_concentration * last_concentration * curr_time;
       },
       5);
   regulate_example.AddGene(
-      [](double curr_time, double last_concentration) {
+      [](float curr_time, float last_concentration) {
         return last_concentration + curr_time + 3;
       },
       7);
 
   // 4. Define initial model -- in this example just one cell.
-  auto construct = [&](const std::array<double, 3>& position) {
+  auto construct = [&](const std::array<float, 3>& position) {
     Cell cell(position);
     cell.SetDiameter(30);
     cell.SetAdherence(0.4);
@@ -56,7 +56,7 @@ inline int Simulate(int argc, const char** argv) {
     cell.AddBiologyModule(regulate_example);
     return cell;
   };
-  const std::vector<std::array<double, 3>>& positions = {{0, 0, 0}};
+  const std::vector<std::array<float, 3>>& positions = {{0, 0, 0}};
   ModelInitializer::CreateCells(positions, construct);
 
   // 5. Run simulation
@@ -69,7 +69,7 @@ inline int Simulate(int argc, const char** argv) {
   const auto& concentrations = regulate_genes->GetConcentrations();
   std::cout << "Gene concentrations after " << Param::total_steps_
             << " time steps" << std::endl;
-  for (double concentration : concentrations) {
+  for (float concentration : concentrations) {
     std::cout << concentration << std::endl;
   }
