+++ src/biology_module/regulate_genes.h
@@ -25,18 +25,18 @@ struct RegulateGenes : public BaseBiologyModule {
   /// \param first_derivative differential equation in the form:
   ///        `slope = f(time, last_concentration)` -- e.g.:
   ///
-  ///              [](double time, double last_concentration) {
+  ///              [](float time, float last_concentration) {
   ///                 return 1 - time * last_concentration;
   ///              }
   ///
   /// \param initial_concentration
-  void AddGene(std::function<double(double, double)> first_derivative,
-               double initial_concentration) {
+  void AddGene(std::function<float(float, float)> first_derivative,
+               float initial_concentration) {
     first_derivatives_.push_back(first_derivative);
     concentrations_.push_back(initial_concentration);
   }
 
-  const std::vector<double>& GetConcentrations() const {
+  const std::vector<float>& GetConcentrations() const {
     return concentrations_;
   }
 
@@ -50,22 +50,22 @@ struct RegulateGenes : public BaseBiologyModule {
     if (Param::numerical_ode_solver_ == Param::NumericalODESolver::kEuler) {
       // Euler
       for (uint64_t i = 0; i < first_derivatives_.size(); i++) {
-        double slope = first_derivatives_[i](absolute_time, concentrations_[i]);
+        float slope = first_derivatives_[i](absolute_time, concentrations_[i]);
         concentrations_[i] += slope * timestep;
       }
     } else if (Param::numerical_ode_solver_ ==
                Param::NumericalODESolver::kRK4) {
       // Runge-Kutta 4
       for (uint64_t i = 0; i < first_derivatives_.size(); i++) {
-        double interval_midpoint = absolute_time + timestep / 2.0;
-        double interval_endpoint = absolute_time + timestep;
+        float interval_midpoint = absolute_time + timestep / 2.0;
+        float interval_endpoint = absolute_time + timestep;
 
-        double k1 = first_derivatives_[i](absolute_time, concentrations_[i]);
-        double k2 = first_derivatives_[i](
+        float k1 = first_derivatives_[i](absolute_time, concentrations_[i]);
+        float k2 = first_derivatives_[i](
             interval_midpoint, concentrations_[i] + timestep * k1 / 2.0);
-        double k3 = first_derivatives_[i](
+        float k3 = first_derivatives_[i](
             interval_midpoint, concentrations_[i] + timestep * k2 / 2.0);
-        double k4 = first_derivatives_[i](interval_endpoint,
+        float k4 = first_derivatives_[i](interval_endpoint,
                                           concentrations_[i] + timestep * k3);
 
         concentrations_[i] += timestep / 6.0 * (k1 + 2 * k2 + 2 * k3 + k4);
@@ -75,12 +75,12 @@ struct RegulateGenes : public BaseBiologyModule {
 
  private:
   /// Store the current concentration for each gene
-  std::vector<double> concentrations_ = {};
+  std::vector<float> concentrations_ = {};
 
   /// Store the gene differential equations, which define how the concentration
   /// change.
   /// New functions can be added through method AddGene()
-  std::vector<std::function<double(double, double)>> first_derivatives_ = {};
+  std::vector<std::function<float(float, float)>> first_derivatives_ = {};
 
   ClassDefNV(RegulateGenes, 1);
 };
