+++ src/debug.h
@@ -6,7 +6,7 @@
 #include <iomanip>
 #include <iostream>
 
-inline void Print(const std::array<double, 3>& a, int precision = 10) {
+inline void Print(const std::array<float, 3>& a, int precision = 10) {
   std::cout << std::setprecision(precision) << a[0] << ", " << a[1] << ", "
             << a[2] << std::endl;
 }
